+++ src/Link/index.js
@@ -59,7 +59,7 @@ Link.propTypes = {
     theme: React.PropTypes.string,
     size: React.PropTypes.oneOf(['s', 'm', 'l', 'xl']),
     className: React.PropTypes.string,
-    type: React.PropTypes.type,
+    type: React.PropTypes.string,
     disabled: React.PropTypes.bool,
     focused: React.PropTypes.bool,
     target: React.PropTypes.string,