+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 * Add missing "key" props ([#47](https://github.com/narqo/react-islands/issues/47)).
 * Nothing works in IE 11 ([#45](https://github.com/narqo/react-islands/issues/45)).
 
-## Other
+### Other
 
 * Fix project license.
 * Update readme.